small bug fixes and progress on delete tour

+++ client/src/components/ManageTours/ManageTours.jsx
@@ -1,6 +1,7 @@
-import { Grid } from '@mui/material';
+import { Button, Dialog, DialogActions, DialogTitle, Grid } from '@mui/material';
 import axios from 'axios';
 import { useContext, useEffect, useState } from 'react';
+import { toast } from 'react-toastify';
 import { Context } from '../../context/Context';
 import TourCard from '../shared/TourCard/TourCard';
 import EmptyCard from './EmptyCard/EmptyCard';
@@ -8,7 +9,9 @@ import './_ManageTours.scss';
 
 function ManageTours() {
   const [tourData, setTourData] = useState([]);
+  const [open, setOpen] = useState(false);
   const { user } = useContext(Context);
+  const [isConfirmed, setIsConfirmed] = useState(false);
 
   const loadToursData = async () => {
     const allTours = await axios.get(`http://localhost:4000/api/tours`);
@@ -22,16 +25,32 @@ function ManageTours() {
   //const handleEdit = () => {};
 
   const handleDelete = async (tourId) => {
-    await axios.delete(`http://localhost:4000/api/tours/${tourId}`, {
-      headers: {
-        authorization: `Bearer ${user.token}`,
-      },
-    });
-    const allTours = await axios.get(`http://localhost:4000/api/tours`);
-    if (allTours.status === 200) {
-      setTourData(allTours.data.data);
+    setOpen(true);
+    if (isConfirmed) {
+      await axios.delete(`http://localhost:4000/api/tours/${tourId}`, {
+        headers: {
+          authorization: `Bearer ${user.token}`,
+        },
+      });
+      toast.success('Tour is deleted!');
+      const allTours = await axios.get(`http://localhost:4000/api/tours`);
+      if (allTours.status === 200) {
+        setTourData(allTours.data.data);
+      } else {
+        console.error('Something went wrong');
+      }
+      setIsConfirmed(false);
     } else {
-      console.error('Something went wrong');
+      return;
+    }
+  };
+
+  const handleClose = (answer) => {
+    if (answer === 'yes') {
+      setIsConfirmed(true);
+      setOpen(false);
+    } else {
+      setOpen(false);
     }
   };
 
@@ -46,7 +65,7 @@ function ManageTours() {
         container
         spacing={2}
         direction="row"
-        justifyContent="space-around"
+        justifyContent="flex-start"
         alignItems="center"
         className="manage-tours-wrapper"
       >
@@ -75,6 +94,31 @@ function ManageTours() {
           );
         })}
       </Grid>
+      <Dialog
+        open={open}
+        onClose={handleClose}
+        aria-labelledby="alert-dialog-title"
+        aria-describedby="alert-dialog-description"
+      >
+        <DialogTitle id="alert-dialog-title">{'Delete this tour?'}</DialogTitle>
+        <DialogActions>
+          <Button
+            onClick={() => {
+              handleClose('no');
+            }}
+          >
+            No
+          </Button>
+          <Button
+            onClick={() => {
+              handleClose('yes');
+            }}
+            autoFocus
+          >
+            Yes
+          </Button>
+        </DialogActions>
+      </Dialog>
     </div>
   );
 }
